Create Code_V2.0_Milli_Strand1
Hello guys. This code is restructured with our own for loop. We started from scratch
and have been able to activate a strip of LEDs. If we wire the LEDs in series, then we can 
run multiple strands of LEDs at the same time. This means each rib of our 
umbrella can operate independently.  They will be wired in a row all together,
but will operate "visually" separately. 