Create LEDdrip
This is an Arduino sketch, which drives LED strips with a TM1809 chip. Currently there are about 72 leds on one strip being driven by one pin. In addition, the LEDs are activated by a Maxbotix ultrasonic sensor. I'm using this inside the ribs of an umbrella to simulate dripping rain. Final sketch will have eight sections, which includes one lighted drip for each rib of the umbrella. 