3/16/13 Drip_Millis_For_Arrays
Streamlined code even more. Everything that can be within an array is,
and is linked throughout. Change the strand lengths to whatever you
like, and evertyhing is flexible. Color changed per strand to see the
effect. I can't quite get my head wrapped around the array that would
be needed to make each strand automatic. As it is now, each strand will
be added manually.